Removing client package and moving contents back to top level
