+++ src/test/java/io/pravega/dataimporter/util/PravegaStreamMirroringJobTest.java
@@ -8,7 +8,7 @@ import io.pravega.client.stream.impl.JavaSerializer;
 import io.pravega.connectors.flink.FlinkPravegaReader;
 import io.pravega.connectors.flink.FlinkPravegaWriter;
 import io.pravega.connectors.flink.PravegaWriterMode;
-import io.pravega.dataimporter.client.AppConfiguration;
+import io.pravega.dataimporter.AppConfiguration;
 import io.pravega.dataimporter.jobs.AbstractJob;
 import io.pravega.dataimporter.jobs.PravegaStreamMirroringJob;
 import io.pravega.dataimporter.utils.PravegaRecord;
